feat/connection: Start TCP listener using TcpMappingSocket
